Move to using IFabricJavadocProvider Co-authored-by: mentlerd <mentlerd@gmail.com>
