+++ build.gradle
@@ -50,7 +50,7 @@ dependencies {
 
 	// decompilers
 	implementation ('net.fabricmc:procyon-fabric-compilertools:0.5.35.+')
-	implementation ('org.jetbrains:intellij-fernflower:1.0.0.9')
+	implementation ('org.jetbrains:intellij-fernflower:1.1.0.11')
 
 	// source code remapping
 	implementation ('org.cadixdev:mercury:0.1.0.fabric-SNAPSHOT')
